Rename rads.config._builders to rads.config.builders.
